notebooks/samples/tensorflow/sentiment_analysis/dataflow/requirements.txt
apache-beam[gcp]==2.18.0
cython
google-api-python-client
google-cloud-language
google-api-core
